+++ db/migrate/20150728173211_add_child_id_to_answer_session.rb
@@ -1,0 +1,6 @@
+class AddChildIdToAnswerSession < ActiveRecord::Migration
+  def change
+    add_column :answer_sessions, :child_id, :integer
+    add_index :answer_sessions, :child_id
+  end
+end